requirements.txt
mcdreforged>=2.0.0a1
pycryptodome
pyrebase4